0-print_list.c

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -10,7 +10,7 @@
 
 size_t print_list(const list_t *h)
 {
-	size_t count_elements = 0;
+	unsigned int  count_elements = 0;
 
 	while (h != NULL)
 	{
@@ -18,6 +18,7 @@ size_t print_list(const list_t *h)
 			printf("[0] (nil)\n");
 		else
 			printf("[%d] %s\n", h->len, h->str);
+
 		h = h->next;
 		count_elements++;
 	}
